[#1542] Run documentation search tests explicitly in docs CI
Move Node.js search-ranking checks outside default pytest collection.
Install Node.js in the docs action and fail CI on test failures or skips.
